requirements.txt
requests
python-dotenv
fastapi
uvicorn[standard]
python-binance
matplotlib
